FIRST AI
- create a Basic_solver class that will be used in the MineSweeper class
to slove the game

the solver is very basic but can solve most easy games and sometimes
might get lucky and solve a medium game

there is definitely more room for improvments in many areas for example:

-the class contains one big method called "makeGuess" that
returns a list of possible locations to be opened but we only use
one location each time then we regenerate the whole list again.

-the "makeRandomGuess" method doesn't return a random location
but rather the first possible location starting from the top left
corner of the field.